xdg: move xdg.h to include
